packages/backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
python-jose[cryptography]>=3.3
authlib>=1.3
httpx>=0.27
pydantic>=2.0
pydantic-settings>=2.0
pytest>=8.0
pytest-asyncio>=0.23
respx>=0.21
